algorithmServer/requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
